RH6: hv_netvsc: fix RSS hashing issues
In this patch, several issues are addressed:
1. move netvsc_set_hash() and related functions to hv_compat.h, to keep
lis-next files as close to upstream as possible;
2. in netvsc_set_hash(), don't hash DHCP Discovery message packets, as
in rhel6.x 32 bits those packets may have transport_header point to
invalid memory address;
3. in netvsc_set_hash(), check transport_header pointer for all
ipv4/ipv6
tcp/udp packets and don't hash those packets if their transport_header
point to NULL.
